+++ src/main/java/com/etz/libraryapi/services/CatalogService.java
@@ -3,15 +3,20 @@ package com.etz.libraryapi.services;
 import com.etz.libraryapi.config.Mapper;
 import com.etz.libraryapi.domains.requests.CatalogRequest;
 import com.etz.libraryapi.domains.requests.CreateNewCatalogRequest;
+import com.etz.libraryapi.domains.requests.GenericDeleteRequest;
 import com.etz.libraryapi.domains.responses.AppResponse;
+import com.etz.libraryapi.domains.responses.CatalogResponse;
 import com.etz.libraryapi.models.Catalog;
+import com.etz.libraryapi.models.Librarian;
 import com.etz.libraryapi.repositories.CatalogueRepo;
+import com.etz.libraryapi.repositories.LibrarianRepo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,43 +25,62 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class CatalogService {
     private final CatalogueRepo catalogRepo;
-    private  Mapper mapper;
+    private final LibrarianRepo librarianRepo;
+    private final Mapper mapper;
 
-    public ResponseEntity<AppResponse<List<Catalog>>> getCatalogs() {
+    public ResponseEntity<AppResponse<List<CatalogResponse>>> getCatalogs() {
         List<Catalog> isCatalog = catalogRepo.findAll();
-//        List<CatalogResponse> catalogResponses = isCatalog.stream().map(catalog -> mapper.modelMapper().map(catalog, CatalogResponse.class)).collect(Collectors.toList()); // getting null pointer exception
+        ArrayList<CatalogResponse> catalogResponses = new ArrayList<>();
         if (isCatalog.size() != 0) {
-            return new ResponseEntity<>(new AppResponse<>(true, isCatalog), HttpStatus.FOUND);
+            for (Catalog catalog : isCatalog) {
+                CatalogResponse response = mapper.modelMapper().map(catalog, CatalogResponse.class);
+                response.setBooks(catalog.getCatalogBooks());
+                catalogResponses.add(response);
+            }
+            return new ResponseEntity<>(new AppResponse<>(true, catalogResponses), HttpStatus.FOUND);
         }
         throw new IllegalStateException("Catalog list is empty");
     }
 
-    public ResponseEntity<AppResponse<String>> createNewCatalog(CreateNewCatalogRequest catalogs) {
-        try {
-            for (String category : catalogs.getCatalogNames()) {
-                Catalog catalog = new Catalog();
-                catalog.setName(category.toUpperCase());
-                catalogRepo.save(catalog);
-            }
-            return new ResponseEntity<>(new AppResponse<>(true, "Successful"), HttpStatus.CREATED);
+    public ResponseEntity<AppResponse<String>> createNewCatalog(CreateNewCatalogRequest request) {
+        Optional<Librarian> librarian = librarianRepo.findById(request.getLibrarianId());
+        if (librarian.isPresent()) {
+            try {
+                for (String category : request.getCatalogNames()) {
+                    Catalog catalog = new Catalog();
+                    catalog.setName(category.toUpperCase());
+                    catalogRepo.save(catalog);
+                }
+                return new ResponseEntity<>(new AppResponse<>(true, "Successful"), HttpStatus.CREATED);
 
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
         }
+        throw new IllegalStateException("Unauthorized");
 
     }
 
     public ResponseEntity<AppResponse<String>> editCatalog(Long id, CatalogRequest request) {
-        Catalog catalog = catalogRepo.findById(id).orElseThrow(()->new IllegalArgumentException("Catalog does not exist"));
+        Optional<Librarian> librarian = librarianRepo.findById(request.getLibrarianId());
+        if (librarian.isPresent()) {
+            Catalog catalog = catalogRepo.findById(id).orElseThrow(() -> new IllegalArgumentException("Catalog does not exist"));
             catalog.setName(request.getName().toUpperCase());
             catalogRepo.save(catalog);
             return new ResponseEntity<>(new AppResponse<>(true, "Successful"), HttpStatus.OK);
+        }
+        throw new IllegalStateException("Unauthorized");
+
     }
 
-    public ResponseEntity<AppResponse<?>> deleteCatalog(Long id) {
-        Catalog catalog = catalogRepo.findById(id).orElseThrow(()-> new IllegalArgumentException("Catalog does not exist"));
-        log.info("Optional isCatalog: {}", catalog);
+    public ResponseEntity<AppResponse<?>> deleteCatalog(Long id, GenericDeleteRequest request) {
+        Optional<Librarian> librarian = librarianRepo.findById(request.getLibrarianId());
+        if (librarian.isPresent()) {
+            Catalog catalog = catalogRepo.findById(id).orElseThrow(() -> new IllegalArgumentException("Catalog does not exist"));
+            log.info("Optional isCatalog: {}", catalog);
             catalogRepo.delete(catalog);
             return new ResponseEntity<>(new AppResponse<>(true, ""), HttpStatus.NO_CONTENT);
         }
+        throw new IllegalStateException("Unauthorized");
+    }
 }
